Resolve code review comments

+++ gloo/allreduce_ring_chunked.h
@@ -30,7 +30,7 @@ class AllreduceRingChunked : public Algorithm {
         bytes_(count_ * sizeof(T)),
         fn_(fn) {
     // Use chunks of no less than 1024 bytes (256 * sizeof(float))
-    constexpr unsigned long minSize = 256;
+    constexpr unsigned long long minSize = 256;
     chunks_ = this->contextSize_ * 2;
 #ifdef _WIN32
   chunkSize_ = std::max((size_t)minSize, (size_t)((count_ + chunks_ - 1) / chunks_));
